generic tools, nav drawers, chat init

+++ src/com/larswerkman/holocolorpicker/ColorPicker.java
@@ -35,7 +35,7 @@ import com.abs192.codraw.R;
 
 /**
  * Displays a holo-themed color picker.
- * 
+ *
  * <p>
  * Use {@link #getColor()} to retrieve the selected color. <br>
  * Use {@link #addSVBar(SVBar)} to add a Saturation/Value Bar. <br>
@@ -119,7 +119,7 @@ public class ColorPicker extends View {
 	/**
 	 * {@code true} if the user clicked on the pointer to start the move mode. <br>
 	 * {@code false} once the user stops touching the screen.
-	 * 
+	 *
 	 * @see #onTouchEvent(android.view.MotionEvent)
 	 */
 	private boolean mUserIsMovingPointer = false;
@@ -146,18 +146,18 @@ public class ColorPicker extends View {
 
 	/**
 	 * Number of pixels the origin of this view is moved in X- and Y-direction.
-	 * 
+	 *
 	 * <p>
 	 * We use the center of this (quadratic) View as origin of our internal
 	 * coordinate system. Android uses the upper left corner as origin for the
 	 * View-specific coordinate system. So this is the value we use to translate
 	 * from one coordinate system to the other.
 	 * </p>
-	 * 
+	 *
 	 * <p>
 	 * Note: (Re)calculated in {@link #onMeasure(int, int)}.
 	 * </p>
-	 * 
+	 *
 	 * @see #onDraw(android.graphics.Canvas)
 	 */
 	private float mTranslationOffset;
@@ -256,9 +256,9 @@ public class ColorPicker extends View {
 	/**
 	 * An interface that is called whenever the color is changed. Currently it
 	 * is always called when the color is changes.
-	 * 
+	 *
 	 * @author lars
-	 * 
+	 *
 	 */
 	public interface OnColorChangedListener {
 		public void onColorChanged(int color);
@@ -267,7 +267,7 @@ public class ColorPicker extends View {
 	/**
 	 * An interface that is called whenever a new color has been selected.
 	 * Currently it is always called when the color wheel has been released.
-	 * 
+	 *
 	 */
 	public interface OnColorSelectedListener {
 		public void onColorSelected(int color);
@@ -275,7 +275,7 @@ public class ColorPicker extends View {
 
 	/**
 	 * Set a onColorChangedListener
-	 * 
+	 *
 	 * @param {@code OnColorChangedListener}
 	 */
 	public void setOnColorChangedListener(OnColorChangedListener listener) {
@@ -284,7 +284,7 @@ public class ColorPicker extends View {
 
 	/**
 	 * Gets the onColorChangedListener
-	 * 
+	 *
 	 * @return {@code OnColorChangedListener}
 	 */
 	public OnColorChangedListener getOnColorChangedListener() {
@@ -293,7 +293,7 @@ public class ColorPicker extends View {
 
 	/**
 	 * Set a onColorSelectedListener
-	 * 
+	 *
 	 * @param {@code OnColorSelectedListener}
 	 */
 	public void setOnColorSelectedListener(OnColorSelectedListener listener) {
@@ -302,7 +302,7 @@ public class ColorPicker extends View {
 
 	/**
 	 * Gets the onColorSelectedListener
-	 * 
+	 *
 	 * @return {@code OnColorSelectedListener}
 	 */
 	public OnColorSelectedListener getOnColorSelectedListener() {
@@ -455,8 +455,8 @@ public class ColorPicker extends View {
 		mColorWheelRectangle.set(-mColorWheelRadius, -mColorWheelRadius,
 				mColorWheelRadius, mColorWheelRadius);
 
-		mColorCenterRadius = (int) ((float) mPreferredColorCenterRadius * ((float) mColorWheelRadius / (float) mPreferredColorWheelRadius));
-		mColorCenterHaloRadius = (int) ((float) mPreferredColorCenterHaloRadius * ((float) mColorWheelRadius / (float) mPreferredColorWheelRadius));
+		mColorCenterRadius = (int) (mPreferredColorCenterRadius * ((float) mColorWheelRadius / (float) mPreferredColorWheelRadius));
+		mColorCenterHaloRadius = (int) (mPreferredColorCenterHaloRadius * ((float) mColorWheelRadius / (float) mPreferredColorWheelRadius));
 		mCenterRectangle.set(-mColorCenterRadius, -mColorCenterRadius,
 				mColorCenterRadius, mColorCenterRadius);
 	}
@@ -467,10 +467,10 @@ public class ColorPicker extends View {
 
 	/**
 	 * Calculate the color using the supplied angle.
-	 * 
+	 *
 	 * @param angle
 	 *            The selected color's position expressed as angle (in rad).
-	 * 
+	 *
 	 * @return The ARGB value of the color on the color wheel at the specified
 	 *         angle.
 	 */
@@ -506,7 +506,7 @@ public class ColorPicker extends View {
 
 	/**
 	 * Get the currently selected color.
-	 * 
+	 *
 	 * @return The ARGB value of the currently selected color.
 	 */
 	public int getColor() {
@@ -517,7 +517,7 @@ public class ColorPicker extends View {
 	 * Set the color to be highlighted by the pointer. </br> </br> If the
 	 * instances {@code SVBar} and the {@code OpacityBar} aren't null the color
 	 * will also be set to them
-	 * 
+	 *
 	 * @param color
 	 *            The RGB value of the color to highlight. If this is not a
 	 *            color displayed on the color wheel a very simple algorithm is
@@ -571,10 +571,10 @@ public class ColorPicker extends View {
 
 	/**
 	 * Convert a color to an angle.
-	 * 
+	 *
 	 * @param color
 	 *            The RGB value of the color to "find" on the color wheel.
-	 * 
+	 *
 	 * @return The angle (in rad) the "normalized" color is displayed on the
 	 *         color wheel.
 	 */
@@ -686,10 +686,10 @@ public class ColorPicker extends View {
 	/**
 	 * Calculate the pointer's coordinates on the color wheel using the supplied
 	 * angle.
-	 * 
+	 *
 	 * @param angle
 	 *            The position of the pointer expressed as angle (in rad).
-	 * 
+	 *
 	 * @return The coordinates of the pointer's center in our internal
 	 *         coordinate system.
 	 */
@@ -702,7 +702,7 @@ public class ColorPicker extends View {
 
 	/**
 	 * Add a Saturation/Value bar to the color wheel.
-	 * 
+	 *
 	 * @param bar
 	 *            The instance of the Saturation/Value bar.
 	 */
@@ -715,7 +715,7 @@ public class ColorPicker extends View {
 
 	/**
 	 * Add a Opacity bar to the color wheel.
-	 * 
+	 *
 	 * @param bar
 	 *            The instance of the Opacity bar.
 	 */
@@ -740,7 +740,7 @@ public class ColorPicker extends View {
 
 	/**
 	 * Change the color of the center which indicates the new color.
-	 * 
+	 *
 	 * @param color
 	 *            int of the color.
 	 */
@@ -760,7 +760,7 @@ public class ColorPicker extends View {
 
 	/**
 	 * Change the color of the center which indicates the old color.
-	 * 
+	 *
 	 * @param color
 	 *            int of the color.
 	 */
@@ -776,7 +776,7 @@ public class ColorPicker extends View {
 
 	/**
 	 * Set whether the old color is to be shown in the center or not
-	 * 
+	 *
 	 * @param show
 	 *            true if the old color is to be shown, false otherwise
 	 */
@@ -792,7 +792,7 @@ public class ColorPicker extends View {
 	/**
 	 * Used to change the color of the {@code OpacityBar} used by the
 	 * {@code SVBar} if there is an change in color.
-	 * 
+	 *
 	 * @param color
 	 *            int of the color used to change the opacity bar color.
 	 */
@@ -804,7 +804,7 @@ public class ColorPicker extends View {
 
 	/**
 	 * Used to change the color of the {@code SaturationBar}.
-	 * 
+	 *
 	 * @param color
 	 *            int of the color used to change the opacity bar color.
 	 */
@@ -816,7 +816,7 @@ public class ColorPicker extends View {
 
 	/**
 	 * Used to change the color of the {@code ValueBar}.
-	 * 
+	 *
 	 * @param color
 	 *            int of the color used to change the opacity bar color.
 	 */
@@ -828,7 +828,7 @@ public class ColorPicker extends View {
 
 	/**
 	 * Checks if there is an {@code OpacityBar} connected.
-	 * 
+	 *
 	 * @return true or false.
 	 */
 	public boolean hasOpacityBar() {
@@ -837,7 +837,7 @@ public class ColorPicker extends View {
 
 	/**
 	 * Checks if there is a {@code ValueBar} connected.
-	 * 
+	 *
 	 * @return true or false.
 	 */
 	public boolean hasValueBar() {
@@ -846,7 +846,7 @@ public class ColorPicker extends View {
 
 	/**
 	 * Checks if there is a {@code SaturationBar} connected.
-	 * 
+	 *
 	 * @return true or false.
 	 */
 	public boolean hasSaturationBar() {
@@ -855,7 +855,7 @@ public class ColorPicker extends View {
 
 	/**
 	 * Checks if there is a {@code SVBar} connected.
-	 * 
+	 *
 	 * @return true or false.
 	 */
 	public boolean hasSVBar() {
